+++ lib/multi_tag_picker.dart
@@ -1,6 +1,8 @@
-//
-// Copyright 2020 Sarbagya Dhaubanjar
-//
+/// Provides a widget that combines a text field with a row of chips positioned
+/// below the text field. The chips represent values entered by the user by
+/// means of the text field provided by the `flutter_typeahead` library.
+///
+/// See the [FlutterTagging] widget for details.
 library multi_tag_picker;
 
 export 'package:flutter_typeahead/flutter_typeahead.dart';