+++ lib/multi_tag_picker.dart
@@ -8,5 +8,4 @@ library multi_tag_picker;
 export 'package:flutter_typeahead/flutter_typeahead.dart';
 
 export 'src/configurations.dart';
-export 'src/taggable.dart';
 export 'src/tagging.dart';